# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Demo API
  version: 0.2.0
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      security: []
      summary: Login with a real database user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login response
...
  /scheduler/tasks/{id}:
    put:
      summary: Update scheduler task
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete scheduler task
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Deleted
  /scheduler/tasks/{id}/toggle:
    post:
      summary: Toggle scheduler task
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Task enabled state
  /scheduler/tasks/{id}/run:
    post:
      summary: Run scheduler task immediately
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Run result
  /scheduler/tasks/{id}/executions:
    get:
      summary: List scheduler task executions
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Task executions
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password